shipment_ai/requirements.txt
fastapi
uvicorn
sqlalchemy
pdfplumber
google-genai
python-multipart
pydantic
jinja2
